+++ backend/horarios_api/src/main/java/online/horarios_api/course/domain/model/CourseComponent.java
@@ -1,11 +1,12 @@
 package online.horarios_api.course.domain.model;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CourseComponent(
         UUID id,
         String componentType,
-        int weeklyHours,
+        BigDecimal weeklyHours,
         String requiredRoomType,
         int sortOrder,
         boolean isActive
