+++ backend/horarios_api/src/main/java/online/horarios_api/course/domain/model/CourseComponentData.java
@@ -1,8 +1,10 @@
 package online.horarios_api.course.domain.model;
 
+import java.math.BigDecimal;
+
 public record CourseComponentData(
         String componentType,
-        int weeklyHours,
+        BigDecimal weeklyHours,
         String requiredRoomType,
         Integer sortOrder,
         Boolean isActive
