+++ backend/horarios_api/src/main/java/online/horarios_api/course/infrastructure/in/web/dto/CourseResponse.java
@@ -2,6 +2,7 @@ package online.horarios_api.course.infrastructure.in.web.dto;
 
 import online.horarios_api.course.domain.model.Course;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
@@ -13,7 +14,7 @@ public record CourseResponse(
         int cycle,
         int credits,
         int requiredCredits,
-        int weeklyHours,
+        BigDecimal weeklyHours,
         String requiredRoomType,
         boolean isActive,
         List<CourseComponentResponse> components,
